Datetime templatetags split into creations.py and comments.py so they are in their semantic environment(ref #57)
