+++ ddcz/templates/creative-pages/commonarticle-list.html
@@ -2,7 +2,7 @@
 {% load html %}
 {% load creations %}
 {% load pages %}
-{% load datetime %}
+
 
 {% block content %}
 <div id="page-common-article">
