+++ ddcz/templates/creative-pages/item-list.html
@@ -2,7 +2,7 @@
 {% load html %}
 {% load creations %}
 {% load pages %}
-{% load datetime %}
+
 
 {% block content %}
 <div>
