+++ ddcz/templates/creative-pages/monster-list.html
@@ -2,7 +2,7 @@
 {% load html %}
 {% load creations %}
 {% load pages %}
-{% load datetime %}
+
 
 {% block content %}
 <div>
