+++ ddcz/templates/dating/list.html
@@ -2,7 +2,7 @@
 {% load html %}
 {% block content %}
 {% load pages %}
-{% load datetime %}
+
 <div id="page-news">
     <h1 class="page-heading">Seznamka</h1>
     <p>Seznamka slouží k seznamování hráčů a PJů. Pokud chcete něco prodat, nebo něco sháníte vkládejte inzeráty do inzerce.</p>
