+++ ddcz/templates/discussions/phorum-list.html
@@ -1,7 +1,7 @@
 {% extends 'public.html' %}
 {% load html %}
 {% load pages %}
-{% load datetime %}
+{% load comments %}
 {% load users %}
 {% block content %}
 <div id="page-phorum">
