+++ ddcz/templates/links/list.html
@@ -2,7 +2,7 @@
 {% load creations %}
 {% load html %}
 {% load pages %}
-{% load datetime %}
+
 {% block content %}
 <div id="page-links">
     <h1 class="page-heading">Linky</h1>
