+++ ddcz/templates/news/list.html
@@ -1,7 +1,7 @@
 {% extends 'public.html' %}
 {% load html %}
 {% load pages %}
-{% load datetime %}
+{% load creations %}
 
 {% block content %}
 <div id="page-news">
